refactor: move permissions decorator into api utils dir
