+++ src/api/utils/permissions.decorator.ts
@@ -1,11 +1,11 @@
 /*
- * SPDX-FileCopyrightText: 2021 The HedgeDoc developers (see AUTHORS file)
+ * SPDX-FileCopyrightText: 2022 The HedgeDoc developers (see AUTHORS file)
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import { CustomDecorator, SetMetadata } from '@nestjs/common';
 
-import { Permission } from './permissions.enum';
+import { Permission } from '../../permissions/permissions.enum';
 
 /**
  * This decorator gathers the {@link Permission Permission} a user must hold for the {@link PermissionsGuard}
